Task #3326: merge from trunk

+++ branches/CEP-Pipeline-Task3326_standalone/CEP/GSM/bremen/sql/tables/create.table.datasets.sql
@@ -1,5 +1,5 @@
 /**
- * This table contains the information about the dataset that is produced by LOFAR. 
+ * This table contains the information about the dataset that is produced by LOFAR.
  * A dataset has an integration time and consists of multiple frequency layers.
  * taustart_timestamp:  the start time of the integration
  */
@@ -8,8 +8,8 @@ CREATE SEQUENCE "seq_datasets" AS INTEGER;
 CREATE TABLE datasets (
   dsid INT DEFAULT NEXT VALUE FOR "seq_datasets",
   rerun INT NOT NULL DEFAULT '0',
-  dstype TINYINT NOT NULL,
-  process_ts TIMESTAMP NOT NULL, 
+  dstype INT NOT NULL,
+  process_ts TIMESTAMP NOT NULL,
   dsinname VARCHAR(64) NOT NULL,
   dsoutname VARCHAR(64) DEFAULT NULL,
   description VARCHAR(100) DEFAULT NULL,
